backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==2.6.0
Flask-Talisman==1.1.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.26.4
python-dateutil==2.8.2
google-cloud-bigquery==3.17.1
pytz==2024.1
APScheduler==3.10.4
